the Join Table command implement scope Provider but can catch all the table. - I delete on constraint the default scope which was inherited from table. The scope is Still working -change the  return null to return new EmptyScope()
The scope is still working.

Go to The TableTranformation concept which is extend by JoinTables
-change the  return null to return new EmptyScope()

Go to FuturTableRef zhich is a children of joinTable as InputTables
Difined here is own scope. Instead of go look at the ancestor Analysis and build the list of FuturTable I stopped
it at the level of th StatementList.

The next step is to delete all the constraint of futurTableRef and see the impaact of this modification to the scope
and make it inherited from Table